requirements.txt
google-genai>=0.8.0
requests>=2.31.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml>=5.2.0
schedule>=1.2.0
sentence-transformers>=3.0.0
numpy>=1.26.0
